Create takeHomeSimulation3
Takehome Challenge - Sim #3 Implement either a yaw model (turning model) or a motor/torque model. For starting with a motor/torque model, I recommend you research the equation power = torque * angular velcity of the motor which is often represented as P = Torque * RPM * 9000. Use matplotlib animations in the form of the template provided.